not fixed yet

+++ Source/VinkiModSession.cs
@@ -5,5 +5,5 @@ namespace Celeste.Mod.VinkiMod;
 public class VinkiModSession : EverestModuleSession {
     public int[][] sessionArtSpots = [];
     public bool sessionStuffLoaded = false;
-    public float[] vinkiRenderIt = [0,0,0,0,0,0,0];
+    public int[] vinkiRenderIt = [0,0,0,0,0,0,0];
 }
